+++ src/rosros/__init__.py
@@ -13,8 +13,8 @@ Released under the BSD License.
 """
 ## @namespace rosros
 __title__        = "rosros"
-__version__      = "0.2.2.dev8"
-__version_info__ = (0, 2, 2, "dev8")
+__version__      = "0.2.2.dev9"
+__version_info__ = (0, 2, 2, "dev9")
 
 from rosros import api
 from rosros import rclify