+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Changelog for package rosros
 * fix rosros.get_param_names() returning params dictionary instead of list
 * fix rosros.set_param() raising error on no descriptor in ROS2
 * fix util.get_arity() and wrap_arity() raising for built-in functions
+* fix rosros.spin_once() raising error on negative timeout in ROS1
 * fix throttling log messages again
 
 0.2.1 (2023-12-06)